Shelter gallery done
