+++ app/_components/GalleryUpperText/GalleryUpperText.jsx
@@ -5,7 +5,7 @@ const GalleryUpperText = () => {
   return (
     <div className="adoption-banner">
       <div className="banner-content">
-        <h1>Encontre Seu Novo Melhor Amigo</h1>
+        <h1>Encontre o seu novo melhor Amigo</h1>
         <p>Estes adoráveis animais estão à espera de um lar amoroso</p>
         <div className="slogan">
           <span className="heart-icon">♥</span>
