+++ app/_components/ShelterGalleryComponents/ShelterBanner/styles/index.css
@@ -1,0 +1,33 @@
+.shelter-banner {
+  background: linear-gradient(45deg, #4a0e8f, #89229b);
+  color: white;
+  padding: 2rem;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border-radius: 20px;
+}
+
+.shelter-banner h1 {
+  font-size: 2.5rem;
+  margin-bottom: 1rem;
+}
+
+.shelter-banner p {
+  font-size: 1.2rem;
+  margin-bottom: 1.5rem;
+}
+
+.shelter-upper-text h2 {
+  color: #f6ad54;
+  border: none;
+  padding: 0.75rem 1.5rem;
+  border-radius: 5px;
+  transition: background-color 0.3s ease;
+}
+
+.shelter-banner a {
+  color: #f6ad54;
+}
