+++ app/_components/ShelterGalleryComponents/ShelterGalleryCard/style/index.css
@@ -7,10 +7,12 @@
   overflow: hidden;
   transition: transform 0.3s ease;
   margin: 2% 1%;
+  cursor: pointer;
 }
 
 .shelterGallery-card:hover {
   transform: translateY(-5px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.4);
 }
 .shelterGallery-imageDiv {
   height: 80%;
