--- app/backoffice/page.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import "./styles/index.css";
-
-const page = () => {
-  return (
-    <div>
-      <div className="container">
-        <div className="sidebar">FILTROS</div>
-        <div className="banner">FILTROS E BUTTON</div>
-        <div className="box"></div>
-      </div>
-    </div>
-  );
-};
-
-export default page;
