--- app/backoffice/styles/index.css
@@ -1,33 +1,0 @@
-.sidebar {
-  background-color: rgb(49, 243, 110);
-  width: 15%;
-  height: 85vh;
-  color: rgb(59, 175, 138);
-}
-
-.box {
-  background-color: rgb(108, 100, 218);
-  width: 60%;
-  height: 700px;
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-40%, -40%);
-}
-.container {
-  display: flex;
-  width: 100%;
-  height: 80vh;
-  background-color: rgb(255, 255, 255);
-}
-
-.banner {
-  background-color: rgb(188, 82, 33);
-  z-index: 1;
-  width: 60%;
-  height: 100px;
-  position: absolute;
-  left: 50%;
-  top: 20%;
-  transform: translate(-40%, -5%);
-}
